invoice-ingestion/requirements.txt
google-cloud-billing==1.12.1
google-cloud-firestore==2.13.1
google-cloud-bigquery==3.13.0
google-api-core==2.15.0
python-dotenv==1.0.0
python-dateutil==2.8.2
